Move settings from .profile to .shinit
These settings only really make sense for interactive shells.  They are
exported because they need to be visible to programs invoked by the
shell, I don't expect to modify them so that they can be inherited by
interactive subshells, so unconditionally setting them to the defaults
when starting an interactive shell is fine.